Use existing editor w/preview
